requirements.txt
PySide6>=6.7
mss>=9.0.1
pytesseract>=0.3.10
Pillow>=10.3.0
keyboard>=0.13.5
opencv-python>=4.10.0.84
pytest>=8.3.0
